use type inference to refine element types in array accesses
From a Java bytecode array access instruction, one can only infer
the element type if it is primitive (because there is a distinct
variant for each primitive type). For all other cases, we do not
know any better than java.lang.Object.

This patch fixes that by using type inference to infer the element
type and refining it, if possible.